refactor: 클래스 이름 변경
Dae-Hwa/airbnb/#23
