+++ iOS/airbnb/airbnb/AccommodationConditions/CommonModels/Charge.swift
@@ -1,0 +1,14 @@
+//
+//  Charge.swift
+//  airbnb
+//
+//  Created by Song on 2021/05/27.
+//
+
+import Foundation
+
+class Charge {
+    var minimum: Double?
+    var maximum: Double?
+}
+//ConditionManager의 형태를 잡기 위해 임시적으로 구현한 타입입니다
