+++ iOS/airbnb/airbnb/AccommodationConditions/LocationSearch/ViewController/PopularLocationViewController.swift
@@ -112,17 +112,19 @@ extension PopularLocationViewController: UISearchBarDelegate {
 
 extension PopularLocationViewController: SearchResultDelegate {
     
-    func didSelect(result: LocationSearchResult) {
+    func didSelect(result: Location) {
         pushNextViewController(with: result)
         inactiveSearchController()
         unsetCancelBarButton()
     }
     
-    private func pushNextViewController(with result: LocationSearchResult) {
+    private func pushNextViewController(with result: Location) {
         let storyboard = self.storyboard ?? StoryboardFactory.create(.accommodationConditions)
         let nextViewController = ViewControllerFactory.create(from: storyboard, type: CalendarViewController.self)
-        let accommodationConditions = AccommodationConditions(location: result)
+        
+        let accommodationConditions = ConditionManager(location: result)
         nextViewController.viewModel = CalendarViewModel(conditionManager: accommodationConditions)
+        
         self.navigationItem.backButtonTitle = PopularLocationViewModel.ButtonTitle.back
         self.navigationController?.pushViewController(nextViewController, animated: true)
     }
